moved date ontainer functionality to hook
